feat: StudyGroup 테이블 수정  (#119) - row modal 연결점 추가 - 페이지 네이션을 위한 값들 prop
